requirements.txt
beautifulsoup4==4.13.5
currency-codes==23.6.4
langchain==0.3.27
langchain-community==0.3.27
langchain-google-vertexai==2.0.28
streamlit==1.48.1
google-cloud-aiplatform==1.110.0
google-cloud-discoveryengine==0.13.11
google-genai==1.31.0
pandas==2.3.1
pydantic==2.11.7
python-decouple==3.8
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.15.0
